Created the score generating algorithm based on calculation of overlap of the different fields. Used fuzywuzzy library for keyword matching in score generation

+++ profDetailed/models.py
@@ -6,7 +6,7 @@ class profDetailed(models.Model):
     keywords = models.TextField(blank=True,null=True)
     minWorkEx = models.PositiveIntegerField(blank=True,null=True)
     minYearOfStudy = models.PositiveIntegerField(blank=True,null=True)
-    minCgpa = models.PositiveIntegerField(blank=True,null=True)
+    minCgpa = models.FloatField(blank=True,null=True)
     branch = models.TextField(blank=True,null=True)
