+++ studentDetailed/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 class studentDetailed(models.Model):
     username = models.CharField(primary_key=True,max_length=32)
-    cgpa = models.IntegerField(blank=True,null=True)
+    cgpa = models.FloatField(blank=True,null=True)
     yearOfStudy = models.IntegerField(blank=True,null=True)
     branch = models.CharField(max_length=100,blank=True,null=True)
     institute = models.CharField(max_length=250,blank=True,null=True)
